requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
Flask==1.0.2
Flask-API==1.0
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
nose==1.3.7
pipenv==2018.5.18
pkg-resources==0.0.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pycodestyle==2.4.0
pycparser==2.18
PyJWT==1.6.4
pytest==3.6.0
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
